after merge fixes, fix onboarding background
